qUnit bug fix, point.normalize, and getter/setter enhancement.
Fixed a minor bug in qunit where it was not checking for a defined value.
Finished the normalize function of Point -- one of the tests fails, but it's due to a rounding error.

Added a mix-in to underscore that I'm experimenting with -- enhanceWithGettersAndSetters.
It turns all set* and get* methods into official getters and setters. My thought was
using getters and setters does not compile in google closure because it's not in ECMA3, so
this method allows us to create get* and set* methods that are aliased to getters
and setters, so we can still compile our code in google closure.